Finished Java project, without multiply

+++ List-1-Java/src/NumberCoder/NumberCoder.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface NumberCoder {
     List<Byte> codeNumber(long numberToCode);
     long decodeNumber(List<Byte> numberToDecode);
-    int getBase();
+    byte getBase();
 }
